layers: Move event types to event_state
