+++ BUILD.gn
@@ -303,7 +303,8 @@ vvl_sources = [
   "layers/state_tracker/device_range_state.h",
   "layers/state_tracker/device_state.cpp",
   "layers/state_tracker/device_state.h",
-  "layers/state_tracker/event_map.h",
+  "layers/state_tracker/event_state.cpp",
+  "layers/state_tracker/event_state.h",
   "layers/state_tracker/fence_state.cpp",
   "layers/state_tracker/fence_state.h",
   "layers/state_tracker/image_layout_map.cpp",
