+++ layers/core_checks/cc_state_tracker.h
@@ -22,7 +22,7 @@
 #include "state_tracker/state_tracker.h"
 #include "state_tracker/cmd_buffer_state.h"
 #include "state_tracker/queue_state.h"
-#include "state_tracker/event_map.h"
+#include "state_tracker/event_state.h"
 #include "state_tracker/shader_stage_state.h"
 
 class CoreChecks;
